Create Pages Link Table
A section with Section Heading a grid container with 2 column of 1fr. in left container there are pages title and are linked to there respected pages. in right container there are images coming from three sources 1st page metafield, 2nd block setting,3rd a placeholder image if the both 1st and 2nd are blank.

hovering on the page title the image for this page gone visible and all the others images goes hidden.